Move projection classes to parent package
