+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/aws/athena/DateProjectionFactory.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.aws.athena.projection;
+package io.trino.plugin.hive.aws.athena;
 
 import com.google.common.collect.ImmutableSet;
 import io.trino.spi.TrinoException;
@@ -41,7 +41,7 @@ import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.COLU
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.COLUMN_PROJECTION_RANGE;
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.getProjectionPropertyRequiredValue;
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.getProjectionPropertyValue;
-import static io.trino.plugin.hive.aws.athena.projection.Projection.invalidProjectionException;
+import static io.trino.plugin.hive.aws.athena.Projection.invalidProjectionException;
 import static java.lang.String.format;
 import static java.time.temporal.ChronoUnit.DAYS;
 import static java.time.temporal.ChronoUnit.HOURS;
