+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/aws/athena/InjectedProjectionFactory.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.aws.athena.projection;
+package io.trino.plugin.hive.aws.athena;
 
 import io.trino.spi.type.Type;
