+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/aws/athena/IntegerProjectionFactory.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive.aws.athena.projection;
+package io.trino.plugin.hive.aws.athena;
 
 import io.trino.spi.type.BigintType;
 import io.trino.spi.type.IntegerType;
@@ -27,7 +27,7 @@ import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.COLU
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.COLUMN_PROJECTION_RANGE;
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.getProjectionPropertyRequiredValue;
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.getProjectionPropertyValue;
-import static io.trino.plugin.hive.aws.athena.projection.Projection.invalidProjectionException;
+import static io.trino.plugin.hive.aws.athena.Projection.invalidProjectionException;
 import static java.lang.String.format;
 
 public class IntegerProjectionFactory
