+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/aws/athena/PartitionProjectionService.java
@@ -19,13 +19,6 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.inject.Inject;
 import io.trino.plugin.hive.HiveConfig;
-import io.trino.plugin.hive.aws.athena.projection.DateProjectionFactory;
-import io.trino.plugin.hive.aws.athena.projection.EnumProjectionFactory;
-import io.trino.plugin.hive.aws.athena.projection.InjectedProjectionFactory;
-import io.trino.plugin.hive.aws.athena.projection.IntegerProjectionFactory;
-import io.trino.plugin.hive.aws.athena.projection.Projection;
-import io.trino.plugin.hive.aws.athena.projection.ProjectionFactory;
-import io.trino.plugin.hive.aws.athena.projection.ProjectionType;
 import io.trino.plugin.hive.metastore.Column;
 import io.trino.plugin.hive.metastore.Table;
 import io.trino.spi.TrinoException;
@@ -70,11 +63,11 @@ import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.PART
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.PARTITION_PROJECTION_LOCATION_TEMPLATE;
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.PROPERTY_KEY_PREFIX;
 import static io.trino.plugin.hive.aws.athena.PartitionProjectionProperties.getMetastoreProjectionPropertyKey;
-import static io.trino.plugin.hive.aws.athena.projection.Projection.unsupportedProjectionColumnTypeException;
-import static io.trino.plugin.hive.aws.athena.projection.ProjectionType.DATE;
-import static io.trino.plugin.hive.aws.athena.projection.ProjectionType.ENUM;
-import static io.trino.plugin.hive.aws.athena.projection.ProjectionType.INJECTED;
-import static io.trino.plugin.hive.aws.athena.projection.ProjectionType.INTEGER;
+import static io.trino.plugin.hive.aws.athena.Projection.unsupportedProjectionColumnTypeException;
+import static io.trino.plugin.hive.aws.athena.ProjectionType.DATE;
+import static io.trino.plugin.hive.aws.athena.ProjectionType.ENUM;
+import static io.trino.plugin.hive.aws.athena.ProjectionType.INJECTED;
+import static io.trino.plugin.hive.aws.athena.ProjectionType.INTEGER;
 import static io.trino.spi.StandardErrorCode.INVALID_COLUMN_PROPERTY;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
